flink-cdc/cdc-manager/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
Jinja2==3.1.2
gunicorn==21.2.0
